Backward compatibility for a couple attributes.
I made some small tweaks when writing schema to make the schema a valid schema salad schema. Add a test case to ensure that these attributes are handled either in their new or old version, but also that round tripping the document from native and back yields the newer form that validates.

GalaxyWorkflow.name -> GalaxyWorkflow.label because name is not a valid record field in schema salad sadly and because we're inheriting 'label' from Process anyway.

WorkflowStep.outputs -> Workflow.out - can't overlap with 'outputs' defined at the workflow level but this is more inline with 'in' and CWL's Process definition as well.